+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSImagePartialEventArgs.cs
@@ -6,9 +6,9 @@ public class NSImagePartialEventArgs : EventArgs
 {
 	public NSImageRep Rep { get; set; }
 
-	public long Rows { get; set; }
+	public nint Rows { get; set; }
 
-	public NSImagePartialEventArgs(NSImageRep rep, long rows)
+	public NSImagePartialEventArgs(NSImageRep rep, nint rows)
 	{
 		Rep = rep;
 		Rows = rows;
