+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorChildCriterion.cs
@@ -1,5 +1,6 @@
+using System;
 using Foundation;
 
 namespace AppKit;
 
-public delegate NSObject NSRulerEditorChildCriterion(NSRuleEditor editor, long index, NSObject criterion, NSRuleEditorRowType rowType);
+public delegate NSObject NSRulerEditorChildCriterion(NSRuleEditor editor, nint index, NSObject criterion, NSRuleEditorRowType rowType);
