+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorDisplayValue.cs
@@ -1,5 +1,6 @@
+using System;
 using Foundation;
 
 namespace AppKit;
 
-public delegate NSObject NSRulerEditorDisplayValue(NSRuleEditor editor, NSObject criterion, long row);
+public delegate NSObject NSRulerEditorDisplayValue(NSRuleEditor editor, NSObject criterion, nint row);
