+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSRulerEditorPredicateParts.cs
@@ -1,5 +1,6 @@
+using System;
 using Foundation;
 
 namespace AppKit;
 
-public delegate NSDictionary NSRulerEditorPredicateParts(NSRuleEditor editor, NSObject criterion, NSObject value, long row);
+public delegate NSDictionary NSRulerEditorPredicateParts(NSRuleEditor editor, NSObject criterion, NSObject value, nint row);
