+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableReorder.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate bool NSTableReorder(NSTableView tableView, long columnIndex, long newColumnIndex);
+public delegate bool NSTableReorder(NSTableView tableView, nint columnIndex, nint newColumnIndex);
