+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewCell.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate bool NSTableViewCell(NSTableView tableView, NSCell cell, NSTableColumn tableColumn, long row);
+public delegate bool NSTableViewCell(NSTableView tableView, NSCell cell, NSTableColumn tableColumn, nint row);
