+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewCellGetter.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate NSCell NSTableViewCellGetter(NSTableView tableView, NSTableColumn tableColumn, long row);
+public delegate NSCell NSTableViewCellGetter(NSTableView tableView, NSTableColumn tableColumn, nint row);
