+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewColumnRowPredicate.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate bool NSTableViewColumnRowPredicate(NSTableView tableView, NSTableColumn tableColumn, long row);
+public delegate bool NSTableViewColumnRowPredicate(NSTableView tableView, NSTableColumn tableColumn, nint row);
