+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewColumnRowString.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate string NSTableViewColumnRowString(NSTableView tableView, NSTableColumn tableColumn, long row);
+public delegate string NSTableViewColumnRowString(NSTableView tableView, NSTableColumn tableColumn, nint row);
