+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewRowGetter.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate NSTableRowView NSTableViewRowGetter(NSTableView tableView, long row);
+public delegate NSTableRowView NSTableViewRowGetter(NSTableView tableView, nint row);
