+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewRowPredicate.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate bool NSTableViewRowPredicate(NSTableView tableView, long row);
+public delegate bool NSTableViewRowPredicate(NSTableView tableView, nint row);
