+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTableViewViewGetter.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate NSView NSTableViewViewGetter(NSTableView tableView, NSTableColumn tableColumn, long row);
+public delegate NSView NSTableViewViewGetter(NSTableView tableView, NSTableColumn tableColumn, nint row);
