+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewCellPosition.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate string[] NSTextViewCellPosition(NSTextView view, NSTextAttachmentCell forCell, ulong charIndex);
+public delegate string[] NSTextViewCellPosition(NSTextView view, NSTextAttachmentCell forCell, nuint charIndex);
