+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewClickedEventArgs.cs
@@ -9,9 +9,9 @@ public class NSTextViewClickedEventArgs : EventArgs
 
 	public CGRect CellFrame { get; set; }
 
-	public ulong CharIndex { get; set; }
+	public nuint CharIndex { get; set; }
 
-	public NSTextViewClickedEventArgs(NSTextAttachmentCell cell, CGRect cellFrame, ulong charIndex)
+	public NSTextViewClickedEventArgs(NSTextAttachmentCell cell, CGRect cellFrame, nuint charIndex)
 	{
 		Cell = cell;
 		CellFrame = cellFrame;
