+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewDoubleClickEventArgs.cs
@@ -9,9 +9,9 @@ public class NSTextViewDoubleClickEventArgs : EventArgs
 
 	public CGRect CellFrame { get; set; }
 
-	public ulong CharIndex { get; set; }
+	public nuint CharIndex { get; set; }
 
-	public NSTextViewDoubleClickEventArgs(NSTextAttachmentCell cell, CGRect cellFrame, ulong charIndex)
+	public NSTextViewDoubleClickEventArgs(NSTextAttachmentCell cell, CGRect cellFrame, nuint charIndex)
 	{
 		Cell = cell;
 		CellFrame = cellFrame;
