+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewEventMenu.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate NSMenu NSTextViewEventMenu(NSTextView view, NSMenu menu, NSEvent theEvent, ulong charIndex);
+public delegate NSMenu NSTextViewEventMenu(NSTextView view, NSMenu menu, NSEvent theEvent, nuint charIndex);
