+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewLink.cs
@@ -1,5 +1,6 @@
+using System;
 using Foundation;
 
 namespace AppKit;
 
-public delegate bool NSTextViewLink(NSTextView textView, NSObject link, ulong charIndex);
+public delegate bool NSTextViewLink(NSTextView textView, NSObject link, nuint charIndex);
