+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewTextChecked.cs
@@ -1,5 +1,6 @@
+using System;
 using Foundation;
 
 namespace AppKit;
 
-public delegate NSTextCheckingResult[] NSTextViewTextChecked(NSTextView view, NSRange range, NSTextCheckingTypes checkingTypes, NSDictionary options, NSTextCheckingResult[] results, NSOrthography orthography, long wordCount);
+public delegate NSTextCheckingResult[] NSTextViewTextChecked(NSTextView view, NSRange range, NSTextCheckingTypes checkingTypes, NSDictionary options, NSTextCheckingResult[] results, NSOrthography orthography, nint wordCount);
