+++ Source/Platform/Mac/Xamarin.Mac/AppKit/NSTextViewTooltip.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace AppKit;
 
-public delegate string NSTextViewTooltip(NSTextView textView, string tooltip, ulong characterIndex);
+public delegate string NSTextViewTooltip(NSTextView textView, string tooltip, nuint characterIndex);
